+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/product/model/DTO/ProductResponseDTO.java
@@ -10,7 +10,7 @@ public class ProductResponseDTO {
     private Category category;
     private Integer supplierId;
     private String supplierName;
-    private Integer userId;
+    private Long userId;
     private String userName;
     private String productName;
     private Integer expirationDate;
@@ -24,10 +24,10 @@ public class ProductResponseDTO {
 
     public ProductResponseDTO() {}
 
-    public ProductResponseDTO(Integer productId, Category category, Integer supplierId, String supplierName, Integer userId, String userName,
-                                String productName, Integer expirationDate, String storageMethod,
-                                Integer pricePerBox, Integer quantityPerBox, LocalDateTime productCreatedAt,
-                                LocalDateTime productUpdatedAt, LocalDateTime productDeletedAt, Boolean isDeleted) {
+    public ProductResponseDTO(Integer productId, Category category, Integer supplierId, String supplierName, Long userId, String userName,
+                              String productName, Integer expirationDate, String storageMethod,
+                              Integer pricePerBox, Integer quantityPerBox, LocalDateTime productCreatedAt,
+                              LocalDateTime productUpdatedAt, LocalDateTime productDeletedAt, Boolean isDeleted) {
         this.productId = productId;
         this.category = category;
         this.supplierId = supplierId;
@@ -88,11 +88,11 @@ public class ProductResponseDTO {
         this.supplierName = supplierName;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
