+++ src/main/java/com/ohgiraffers/warehousemanagement/wms/product/model/entity/Product.java
@@ -38,7 +38,7 @@ public class Product {
 
     // 담당자 ID
     @Column(name = "user_id", nullable = false)
-    private Integer userId;
+    private Long userId;
 
     // 상품 이름
     @Column(name = "product_name", length = 50, nullable = false)
@@ -87,7 +87,7 @@ public class Product {
 
     public Product() {}
 
-    public Product(Integer productId, Category category, Supplier supplier, Integer userId, String productName, Integer expirationDate, String storageMethod, Integer pricePerBox, Integer quantityPerBox, LocalDateTime productCreatedAt, LocalDateTime productUpdatedAt, LocalDateTime productDeletedAt, Boolean isDeleted) {
+    public Product(Integer productId, Category category, Supplier supplier, Long userId, String productName, Integer expirationDate, String storageMethod, Integer pricePerBox, Integer quantityPerBox, LocalDateTime productCreatedAt, LocalDateTime productUpdatedAt, LocalDateTime productDeletedAt, Boolean isDeleted) {
         this.productId = productId;
         this.category = category;
         this.supplier = supplier;
@@ -129,11 +129,11 @@ public class Product {
         this.supplier = supplier;
     }
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
